Improve console install and restart services appropriately
The current implementation wipes out the novnc/spice directory
every single time the playbook executes. While it does replace
it afterwards it does not fire a notifier to make the service
restart, resulting in the service failing.

This patch makes the install process idempotent - it will only
wipe out the directory if it does not look the way it is
expected to look. It also fires the notifier to restart services
if the git repo changes.

This should ensure that novnc/spice restart properly on upgrades
and that there are no changes/restarts for the service if the git
SHA or software does not change.

Backport note:
The 'Manage LB' handler notification is removed as it is not
present in pike.

Closes-Bug: #1746523
Change-Id: Ic39f884016a6ba0498110267027783001a087a1c
(cherry picked from commit c41103ba702cd6fd85d8e4d4fe86caef81623f44)